yt_assistant_api/requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn==0.34.1
